requirements.txt
docopt==0.6.2
Pillow==3.4.2
